add another working glbook testcase

+++ tests/glbook/Common/esUtil.h
@@ -185,7 +185,7 @@ GLuint ESUTIL_API esLoadProgram ( const char *vertShaderSrc, const char *fragSha
 ///         if it is not NULL ) as a GL_TRIANGLE_STRIP
 //
 int ESUTIL_API esGenSphere ( int numSlices, float radius, GLfloat **vertices, GLfloat **normals, 
-                             GLfloat **texCoords, GLuint **indices );
+                             GLfloat **texCoords, GLushort **indices );
 
 //
 /// \brief Generates geometry for a cube.  Allocates memory for the vertex data and stores
